jenkins: Fix installing plugins after configured
For this, we generate a random password for the jenkins user
and store it in a file, to use it later for every command of jenkins-cli.

We need to try with and without password, because if the auth method is not
set, it will fail if you try with password.

We can't set the auth method before, because jenkins fails to start if
we set the config.xml file before the first run.