echaloasuerte-3: Use also Task ID in the LOGS_SUFFIX variable
Task.Slot will be always 1 if only running one replica, it
will be 1 or 2 when running 2 replicas, etc. This means that
multiple containers (when restarting the service one gets
created before the old is stopped) will share the same
variable and same log file as result.

Adding the Task ID makes this sort of clashing impossible.